console: Refactor connect in device card
