--- pkg/webui/console/containers/device-profile-section/device-card/connect.js
@@ -1,27 +1,0 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { connect } from 'react-redux'
-
-import { selectDeviceModelById } from '@console/store/selectors/device-repository'
-
-const mapStateToProps = (state, props) => {
-  const { brandId, modelId } = props
-
-  return {
-    model: selectDeviceModelById(state, brandId, modelId),
-  }
-}
-
-export default DeviceCard => connect(mapStateToProps)(DeviceCard)
